pyproject.toml
[project]
name = "contextkeeper"
version = "0.1.0"
description = "MCP Server for zero-context-loss AI tool switching"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "pygments>=2.19.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
]
